# src/main/java/ru/relex/cryptocurrencyexchange/exception/responses/ValidationErrorResponse.java
package ru.relex.cryptocurrencyexchange.exception.responses;

import lombok.Getter;
import lombok.RequiredArgsConstructor;

import java.util.List;

@Getter
@RequiredArgsConstructor
public class ValidationErrorResponse {
    private final List<Violation> violations;
}