# src/main/java/ru/relex/cryptocurrencyexchange/exception/responses/Violation.java
package ru.relex.cryptocurrencyexchange.exception.responses;

import lombok.Getter;
import lombok.RequiredArgsConstructor;

@Getter
@RequiredArgsConstructor
public class Violation {
    private final String fieldName;
    private final String message;

}